requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.3.2
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
attrs==25.1.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.8
colorama==0.4.6
cryptography==44.0.2
distro==1.9.0
et_xmlfile==2.0.0
filelock==3.18.0
frozenlist==1.7.0
fsspec==2025.5.1
gitdb==4.0.12
GitPython==3.1.44
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.174.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
griffe==1.7.3
grpcio==1.73.1
grpcio-status==1.71.2
h11==0.16.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.33.0
idna==3.10
importlib_metadata==8.7.0
Jinja2==3.1.5
jiter==0.10.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
litellm==1.72.6.post1
load-dotenv==0.1.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mcp==1.9.4
mdurl==0.1.2
multidict==6.5.0
narwhals==1.29.0
numpy==2.2.3
openai==1.93.0
openai-agents==0.0.19
openpyxl==3.1.5
packaging==24.2
pandas==2.2.3
pillow==11.1.0
propcache==0.3.2
proto-plus==1.26.1
protobuf==5.29.5
pyarrow==19.0.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.7
pydantic-settings==2.9.1
pydantic_core==2.33.2
pydeck==0.9.1
Pygments==2.19.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
pytz==2025.1
PyYAML==6.0.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.4
rich==13.9.4
rpds-py==0.23.1
rsa==4.9.1
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
sse-starlette==2.3.6
starlette==0.47.0
streamlit==1.44.1
tenacity==9.0.0
tiktoken==0.9.0
tokenizers==0.21.1
toml==0.10.2
tornado==6.4.2
tqdm==4.67.1
types-requests==2.32.4.20250611
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.1
uritemplate==4.2.0
urllib3==2.5.0
uv==0.6.6
uvicorn==0.34.3
watchdog==6.0.0
yarl==1.20.1
zipp==3.23.0
